requirements.txt
ultralytics
opencv-python
numpy
filterpy
requests
fastmcp
